pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "PixelPhantomX"
version = "1.0.0"
authors = [
    { name = "Aditya Bhatt", email = "info.adityabhatt3010@gmail.com" }
]
description = "A tool for generating ghost images to confuse AI models."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "numpy",
    "opencv-python",
    "pillow",
    "piexif",
    "pyfiglet",
    "termcolor"
]

[project.urls]
Homepage = "https://github.com/AdityaBhatt3010/PixelPhantomX"
Repository = "https://github.com/AdityaBhatt3010/PixelPhantomX"
Issues = "https://github.com/AdityaBhatt3010/PixelPhantomX/issues"

[project.scripts]
PixelPhantomX = "PixelPhantomX:main"
